refactor: 서버 소켓 연결 부분 파일 분리
