+++ socket/server.js
@@ -1,6 +1,6 @@
 import { createServer } from "http";
 import { Server } from "socket.io";
-import socketHandler from "./socketHandler.js";
+import initSocketConnection from "./loader.js";
 
 const httpServer = createServer();
 const io = new Server(httpServer, {
@@ -8,16 +8,4 @@ const io = new Server(httpServer, {
 });
 const port = 5000;
 
-io.on("connection", socket => {
-  const handler = socketHandler(io, socket);
-
-  socket.on("join", handler.joinChannel);
-  socket.on("leave", handler.leaveChannel);
-  socket.on("chat", handler.sendChatMessage);
-  socket.on("noticeChannelEnded", handler.noticeChannelEnded);
-  socket.on("disconnect", () => {});
-});
-
-httpServer.listen(port, () =>
-  console.log(`Socket.io started on PORT: ${port}`)
-);
+initSocketConnection(io, port);
